requirements.txt
streamlit
numpy
Pillow
keras
ipython